cpp06: fixes

+++ CPP-modules/cpp06/ex01/Serializer.cpp
@@ -1,10 +1,9 @@
 #include "Serializer.hpp"
-#include <cstdint>
 
-uintptr_t Serializer::serialize(Data *ptr) {
-  return reinterpret_cast<uintptr_t>(ptr);
+unsigned long Serializer::serialize(Data *ptr) {
+  return reinterpret_cast<unsigned long>(ptr);
 }
 
-Data *Serializer::deserialize(uintptr_t raw) {
+Data *Serializer::deserialize(unsigned long raw) {
   return reinterpret_cast<Data *>(raw);
 }
